ml-engine/requirements.txt
flask
python-dotenv
google-generativeai
pymongo
Pillow
cryptography
requests
python-jose
uvicorn
kafka-python
google-cloud-storage